refactor : 빈 주입 방식 변경

+++ src/main/java/nextstep/courses/infrastructure/persistence/repository/CourseRepositoryImpl.java
@@ -17,7 +17,7 @@ public class CourseRepositoryImpl implements CourseRepository {
   }
 
   @Override
-  public int save(Course course) {
+  public Long save(Course course) {
     return courseEntityRepository.save(course.toEntity());
   }
